Rename FS and ENV classes
